Add synchronous test episodes; make network architecture clear in config.py; default to non-MuJoCo Pendulum-v0

+++ sac/train.py
@@ -22,12 +22,11 @@ ex.add_config(sac.config.default_config)
 
 
 @ex.capture
-def train_sac(buffer: ReplayBuffer, model: SACModel, train_env,
-              batch_size, n_start_env_steps, log_every_n_steps, checkpoint_every_n_steps,
-              train_n_steps):
-    n_steps = 0
+def train_sac(buffer: ReplayBuffer, model: SACModel, train_env, test_env, n_start_env_steps, log_every_n_steps,
+              checkpoint_every_n_steps, train_n_steps, test_every_n_steps, batch_size, render):
     obs1, done = train_env.reset(), False
     logger = easy_tf_log.Logger(os.path.join(observer.dir, 'sac'))
+    n_steps = 0
     step_rate_measure = RateMeasure(n_steps)
     losses = []
 
@@ -50,11 +49,8 @@ def train_sac(buffer: ReplayBuffer, model: SACModel, train_env,
         loss = model.train(batch)
         losses.append(loss)
 
-        if n_steps % checkpoint_every_n_steps == 0:
-            model.save()
-
         if n_steps % log_every_n_steps == 0:
-            print(f"Trained {n_steps} steps")
+            print(f"Ran {n_steps} steps")
             n_steps_per_second = step_rate_measure.measure(n_steps)
             logger.logkv('dqn/buffer_size', len(buffer))
             logger.logkv('dqn/n_steps', n_steps)
@@ -62,41 +58,58 @@ def train_sac(buffer: ReplayBuffer, model: SACModel, train_env,
             logger.logkv('dqn/loss', np.mean(losses))
             losses = []
 
+        if n_steps % test_every_n_steps == 0:
+            run_test_episodes(model, test_env, n_episodes=10, render=render)
+
+        if n_steps % checkpoint_every_n_steps == 0:
+            model.save()
+
         n_steps += 1
 
 
-def run_test_env(model, model_load_dir, render, env_id, seed, log_dir):
-    env = make_env(env_id, seed, log_dir, 'test')
-    while True:
-        model.load(model_load_dir)
-        obs, done, rewards = env.reset(), False, []
+def run_test_episodes(model, env, n_episodes, render):
+    for _ in range(n_episodes):
+        obs, done, = env.reset(), False
         while not done:
             action = model.step(obs, deterministic=True)
+            obs, _, done, _ = env.step(action)
             if render:
                 env.render()
-            obs, reward, done, info = env.step(action)
+
+
+def run_test_env(model, model_load_dir, render, env_id, seed, log_dir):
+    env = make_env(env_id, seed, log_dir, 'test_async', record_videos=True)
+    while True:
+        model.load(model_load_dir)
+        run_test_episodes(model, env, n_episodes=1, render=render)
 
 
 @ex.automain
-def main(gamma, buffer_size, lr, render, seed, env_id, polyak_coef, temperature, policy_std_min, policy_std_max):
-    env = make_env(env_id, seed, observer.dir, 'train')
+def main(gamma, buffer_size, lr, render, seed, env_id, polyak_coef, temperature, policy_std_min, policy_std_max,
+         async_test, network):
+    train_env = make_env(env_id, seed, observer.dir, 'train', record_videos=False)
+    test_env = make_env(env_id, seed, observer.dir, 'test_sync', record_videos=True)
 
-    buffer = ReplayBuffer(env.observation_space.shape, env.action_space.shape, max_size=buffer_size)
-    obs_dim = env.observation_space.shape[0]
-    n_actions = env.action_space.shape[0]
-    act_lim = env.action_space.high
+    buffer = ReplayBuffer(train_env.observation_space.shape, train_env.action_space.shape, max_size=buffer_size)
+    obs_dim = train_env.observation_space.shape[0]
+    n_actions = train_env.action_space.shape[0]
+    act_lim = train_env.action_space.high
     ckpt_dir = os.path.join(observer.dir, 'checkpoints')
     model = SACModel(obs_dim=obs_dim, n_actions=n_actions, act_lim=act_lim, save_dir=ckpt_dir,
                      discount=gamma, lr=lr, seed=seed, polyak_coef=polyak_coef, temperature=temperature,
-                     std_min_max=(policy_std_min, policy_std_max))
+                     std_min_max=(policy_std_min, policy_std_max), network=network)
     model.save()
 
-    ctx = multiprocessing.get_context('spawn')
-    test_env_proc = ctx.Process(target=run_test_env, daemon=True,
-                                args=(model, ckpt_dir, render, env_id, seed, observer.dir))
-    test_env_proc.start()
+    if async_test:
+        ctx = multiprocessing.get_context('spawn')
+        test_env_proc = ctx.Process(target=run_test_env, daemon=True,
+                                    args=(model, ckpt_dir, render, env_id, seed, observer.dir))
+        test_env_proc.start()
+    else:
+        test_env_proc = None
 
-    train_sac(buffer, model, env)
+    train_sac(buffer, model, train_env, test_env)
 
-    test_env_proc.terminate()
+    if test_env_proc is not None:
+        test_env_proc.terminate()
     return model
